refactor part 2, menus and voice model controls
