+++ main/src/botActions.js
@@ -1,10 +1,15 @@
-import { showMenu, showSettings, showCurrentSettings, loadSettings, showAICoverSettings, showCreateMenu, showEffectsSettings, getRandomMaleVoice, getRandomFemaleVoice } from "./botFunction.js";
+import { loadSettings,getRandomMaleVoice, getRandomFemaleVoice } from "./botFunction.js";
+
 import { MALE_VOICES, FEMALE_VOICES, characters } from "./variables.js";
 import { Markup } from "telegraf";
 import path from "path";
 import fs from 'fs'
-import { getRandomFileContent } from "./functions.js";
 import { registerAdminBotActions } from "./admin/botActions.js";
+import { registerCreateMenuBotActions } from "./createModel/botActions.js";
+import { showCurrentSettings, showMenu } from "./menus/mainMenu.js";
+import { showSettings } from "./menus/settingsMenu.js";
+import { showAICoverSettings } from "./menus/aicoverMenu.js";
+import { showEffectsSettings } from "./menus/effectsMenu.js";
 
 export function groupCharactersByCategory(characters) {
   const categories = {};
@@ -23,6 +28,7 @@ const groupedCharacters = groupCharactersByCategory(characters);
 export function registerBotActions(bot) {
 
     registerAdminBotActions(bot)
+    registerCreateMenuBotActions(bot)
 
     bot.action("effects_settings", async(ctx) =>{
 
@@ -214,92 +220,7 @@ bot.action("toggle_audio_effects", async(ctx) => {
   }
   });
 
-  bot.action("create_voice_info",async (ctx) => {
-    await ctx.reply(`Итак перед тем как дать инструкцию о том что нужно сделать что бы получить свою языковую модель, хочу уточнить пару деталей \nАвтоматического создания модели нет и все образцы которые вы запишете через эти кнопки, я смогу обучить через какое-то время, я сделаю оповещение когда модель будет готова.`)
-    await ctx.reply("\nПеред тем как кидать голосовое сообщение убедитесь что вы находитесь в тихом помещение, нет посторонних шумов, качество модели напрямую зависит от качества записанного голосового")
-    await ctx.reply("\nВы можете выбрать получить текст и перед вами появится 1 из 10 подготовленных текстов, не обязательно зачитывать именно его, вы можете сами зачитать или напеть что угодно, общая продолжительность должна быть более 90 секунд, иначе не ручаюсь за похожесть, чем больше тем лучше, но не более 5 минут.", Markup.inlineKeyboard([
-      Markup.button.callback('Меню', 'show_create_voice_menu')
-    ]))
-  })
-
-  bot.action("create_voice_name",(ctx) =>{
-    ctx.reply("Введите название для вашей голосовой модели")
-    ctx.session.waitForModelName = true
-  })
-
-  bot.action("create_voice_random_text",async (ctx) =>{
-    const text = await getRandomFileContent();
-
-    await ctx.reply(text, Markup.inlineKeyboard([
-      [Markup.button.callback('Записать голосовое', 'create_voice_add_sample'),Markup.button.callback('Назад', 'show_create_voice_menu')]
-    ]))
-  })
-
-  bot.action("create_voice_add_sample",async(ctx) => {
-    if(!ctx.session.voiceModelName) {
-      ctx.reply("Перед тем как скидывать, инициализируйте модель", Markup.inlineKeyboard([
-        Markup.button.callback('Инициализировать модель', 'create_voice_name')
-      ]))
-      return
-    }
-    ctx.reply("Вы можете переслать любое голосовое сообщение от кого-либо боту\nСкиньте аудиофайл ( до 20 мб ) или запишите голосовое")
-    ctx.session.waitForVoice = true
-  })
-
-  bot.action("create_voice_end",async(ctx) => {
-    if(!ctx.session.voiceModelName) {
-      ctx.reply("Перед заканчивать создание, вы должны инициализировать модель", Markup.inlineKeyboard([
-        Markup.button.callback('Инициализировать модель', 'create_voice_name')
-      ]))
-      return
-    }
-    await ctx.reply("Все ваши данные были записанны, ждите когда у админа появится время и он сможет обработать и создать вашу голосовую модель", Markup.inlineKeyboard([
-      Markup.button.callback('Меню', 'menu')
-    ]))
-
-    const uniqueId = ctx.from.id; // получаем уникальный идентификатор пользователя
-    const folderName = ctx.session.voiceModelName
-    let voicePath = `train_voice/${uniqueId}/${folderName}`;
-
-    // создаем папку сессии, если она еще не существует
-    if (!fs.existsSync(voicePath)) {
-      fs.mkdirSync(voicePath, { recursive: true });
-    }
-
-    // создаем и записываем файл
-    const filePath = path.join(voicePath, 'desc.txt');
-    const fileContent = `${ctx.session.voiceModelName}\n${ctx.session.voiceModelDesc}\n${ctx.session.voiceModelGender}`;
-    fs.writeFileSync(filePath, fileContent);
-
-    console.log("Пользователь завершил сбор датасета и ждет когда его модель будет готова")
-    
-    // создаем объект для JSON файла
-    const jsonContent = {
-      username: ctx.from.username,
-      id: ctx.from.id,
-      modelName: ctx.session.voiceModelName,
-      date: new Date().toISOString()
-    };
-
-    // создаем и записываем JSON файл
-    const jsonFilePath = path.join('waitForModel.json');
-    fs.writeFileSync(jsonFilePath, JSON.stringify(jsonContent, null, 2));
-
-    ctx.session.voiceModelName = ""
-    ctx.session.voiceModelDesc = ""
-    ctx.session.voiceModelGender = ""
-  })
-
   
-
-
-  bot.action("show_create_voice_menu",(ctx) =>{
-    ctx.session.waitingForCover = false;
-    ctx.session.mergeAudio = false;
-
-    showCreateMenu(ctx)
-    return
-  })
 
   // Обработчик для кнопки "delete_preset"
   bot.action(/delete_preset:(.+)/, (ctx) => {
