+++ main/src/botFunction.js
@@ -687,71 +687,6 @@ export function saveSuggestion(username, suggestion) {
   });
 }
 
-export const showMenuBtn = async (ctx) => {
-  const keyboard = Markup.inlineKeyboard([
-    Markup.button.callback("Меню", "menu"),
-  ]).resize();
-
-  const message = await ctx.reply('...', keyboard);
-  await ctx.replyWithMarkdown("ㅤ", keyboard);
-};
-
-export const showMenu = async (ctx) => {
-  const message = `*Меню*\n\n\Ваш текущий персонаж: *${ctx.session.name}*\n\Пресет голоса: *${ctx.session.voice_preset}*\nPich: ${ctx.session.pith}\nВы можете сделать предложение по улучшению функционала бота, для этого выберите специальный пункт в меню\n\nВы можете прислать текст и он будет озвучен голосом персонажа ( голос будет роботизированным )\n\nОтправьте голосовое или перешлите уже сделанное, так же вы можете кинуть аудиофайл`;
-
-  const keyboard = Markup.inlineKeyboard([
-    [Markup.button.callback(`Выбрать персонажа`, "characters"),Markup.button.callback(`Показать текущие настройки`, "current_settings")],
-    [Markup.button.callback(`Сделать AI кавер`, "cover"), Markup.button.callback(`Разделить аудио`, "separate") ],
-    [Markup.button.callback("Меню создания голосовых моделей", "show_create_voice_menu")],
-    [Markup.button.callback("Настройки голоса", "settings"), Markup.button.callback(`Настройка AI Кавера`, "aisettings")],
-    [Markup.button.callback(`Сохранить настройки`, "save_preset"), Markup.button.callback(`Загрузить настройки`, "load_preset")],
-    [Markup.button.callback(`Предложения по улучшению бота`, "make_predlog"),Markup.button.callback(`Поддержать автора`, "support")],
-  ]).resize();
-
-  await ctx.replyWithMarkdown(message, keyboard);
-  return message
-};
-
-export async function showCreateMenu(ctx){
-  const message = `Меню создания голосовой модели\nПеред началом инициализируйте модель\nПродолжительность вашего набора данных должна быть 90 секунд или более\n\nНазвание вашей модели:  ${ctx.session.voiceModelName ? ctx.session.voiceModelName :  "Ещё не названно"}\nОписание вашей модели:  ${ctx.session.voiceModelDesc ? ctx.session.voiceModelDesc :  "Не указанно"}\nПол вашей модели:  ${ctx.session.voiceModelGender ? ctx.session.voiceModelGender :  "Не указанно"}`
-
-  const keyboard = Markup.inlineKeyboard([
-    [Markup.button.callback(`Инициализоровать голосовую модель`, "create_voice_name")],
-    [Markup.button.callback(`Добавить голосовой или аудиообразец голоса`, "create_voice_add_sample")],
-    [Markup.button.callback(`Получить текст для озвучивания`, "create_voice_random_text")],
-    [Markup.button.callback(`Информация о создание и сроках`, "create_voice_info")],
-    [Markup.button.callback("Завершить процедуру для создание модели","create_voice_end")],
-    [Markup.button.callback(`Назад`, "menu")],
-  ]).resize();
-
-  await ctx.replyWithMarkdown(message, keyboard);
-  return message
-}
-
-export async function showCurrentSettings(ctx) {
-  const session = ctx.session;
-
-  const photo = { source: fs.readFileSync(session.char_photo) };
-  const newMessage = await ctx.replyWithPhoto(photo);
-
-  const message = `Текущие настройки сессии: ㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤㅤ\n` +
-    `Высота тона: ${session.pith}\n` +
-    `Метод: ${session.method}\n` +
-    `Размер mangio-crepe hop: ${session.mangio_crepe_hop}\n` +
-    `Сила влияния индекса: ${session.feature_ratio}\n` +
-    `Защита голоса: ${session.protect_voiceless}\n` +
-    `Имя: ${session.name}\n` +
-    `Ваш пол: ${session.voice_preset} \n` +
-    `Голос текстовой модели: ${session.voiceActor}`;
-
-  const keyboard = Markup.inlineKeyboard([
-    Markup.button.callback("Назад", "menu"),
-  ])
-
-  await ctx.replyWithMarkdown(message, keyboard);
-  await deletePreviousMessage(ctx);
-};
-
 export function noteOctaveToFrequency(note, octave) {
   var noteMap = {
       'C': -9,
@@ -787,82 +722,4 @@ export function noteOctaveToFrequency(note, octave) {
   }
 }
 
-export async function showSettings(ctx) {
-  
-  const session = ctx.session;
 
-  const pithDescription = "Pith - вв тона, если выбираешь женскую модель но ты мужчина ставь 12, если наоборот -12, если пол совпадает то 0";
-  const methodDescription = "Method - Метод обработки голоса: harvest - лучше выделяет особенности голоса , crepe - более медленный, но лучше обрабатывает голос в целом, mango-crepe - Улучшенный вариант creep с возможностью выбрать Mangio-Crepe Hop";
-  const mangioCrepeHopDescription = "Mangio-Crepe Hop - означает время, необходимое говорящему для перехода на резкую высоту тона. При меньшей длине скачка требуется больше времени, чтобы сделать вывод, но он более точен. Оптимальное значение 128";
-  const featureRatioDescription = "Feature ratio - На сколько голос будет подкоректирован согласно речевым особенностям модели, может вызвать эффект метала";
-  const protectVoicelessDescription = "Protect voiceless - Защита безголосых согласных и звуков дыхания для предотвращения артефактов музыки. Установите значение0,5 для отключения. Уменьшите значение для усиления защиты, но это может снизить точность индексирования";
-  const voiceActorDescription = "Модель для голоса - Изначальный голос для преобразования модели из текста в речь"
-  const outputType = "Тип ответа - Тип присылаемого аудио в ответ на голосовое, 2 режима - audio и voice"
-
-  const autoTuneDesc = `Автотюн, который регулируется нейросетью, конфликтует с автотюном в эффектах, так что может работать только 1` 
-  const aboutMinMax = `Min-Max Pitch, определяет допустимый диапазон частоты (pitch), применяемый в методе rmvpe+.\nДопустимые значения в Гц варьируются от 1 до 16000.Также вы можете использовать музыкальные ноты для определения диапазона.Например, 'C0' соответствует частоте примерно 16 Гц, а 'E10' соответствует частоте примерно 15600 Гц.` 
-
-  const settingsMessage = [pithDescription, aboutMinMax, methodDescription, mangioCrepeHopDescription, featureRatioDescription, protectVoicelessDescription,autoTuneDesc, voiceActorDescription,outputType].join("\n\n");
-
-  const settingsKeyboard = Markup.inlineKeyboard([
-    [Markup.button.callback(`Pith: ${session.pith}`, "set_pith"), Markup.button.callback(`Method: ${session.method}`, "set_method")],
-    [Markup.button.callback(`Mangio-Crepe Hop: ${session.mangio_crepe_hop}`, "set_mangio_crepe_hop"), Markup.button.callback(`Feature Ratio: ${session.feature_ratio}`, "set_feature_ratio")],
-    [Markup.button.callback(`Protect Voiceless: ${session.protect_voiceless}`, "set_protect_voiceless"),Markup.button.callback(`Изменить скорость речи ${session.voice_speed}`, "set_voice_speed")],
-    [Markup.button.callback(`Min Pith: ${session.minPich}`, "set_minPich"),Markup.button.callback(`Max Pith ${session.maxPich}`, "set_maxPich")],
-    [Markup.button.callback(`Autotune: ${session.neuroAutoTune}`,"toggle_autotune")],
-    [Markup.button.callback(`Изменить текстовую голосовую модель ${session.voiceActor}`, "set_voice")],
-    [Markup.button.callback(`Тип ответа на голосовое: ${session.voiceOrAudioOut}`, "set_out_voice")],
-    [Markup.button.callback("Меню", "menu")],
-  ]).resize();
-
-  await ctx.replyWithMarkdown(settingsMessage, settingsKeyboard);
-  await deletePreviousMessage(ctx);
-}
-
-export async function showAICoverSettings(ctx) {
-  
-  const session = ctx.session;
-
-  const mainDesk = "Тут вы увидете настройки которые нужны для создания ИИ кавера и разделения аудио на вокал и инструментал по команде /separate"
-  const vocalVolumeDescription = "Громкость вокала - Уровень громкости вокала может быть от 0 до 3";
-  const instrumentalVolumeDescription = "Громкость инструментала - Уровень громкости инструментала может быть от 0 до 3";
-  const audioProcessPowerDescription = "Модификаторы при обработке вокала - есть 3 модификатора, echo - убрать эхо из вокала, backvocal - убрать бэк вокал и both - убрать и то и то\nКаждый модификатор добавляет примерно по дополнительной минуте"
-  const echoDesc = `Добавить эхо, сейчас оно - ${ctx.session.echoOn === true ? "Включенно" : "Выключенно"}`
-  const reverbDesc = `Добавить реверберацию, сейчас она - ${ctx.session.reverbOn === true ? "Включенна" : "Выключенна"}`
-  const autoTuneDesc = `Добавить автотюн, сейчас он - ${ctx.session.autoTune === true ? "Включенн" : "Выключенн"}`
-
-  const settingsMessage = [mainDesk,vocalVolumeDescription, instrumentalVolumeDescription,audioProcessPowerDescription,reverbDesc,echoDesc,autoTuneDesc].join("\n\n");
-
-  const settingsKeyboard = Markup.inlineKeyboard([
-    [Markup.button.callback(`Громкость вокала: ${session.voice_volume}`, "set_vocal_volume")], 
-    [Markup.button.callback(`Громкость инструментала: ${session.instrumnet_volume}`, "set_instrumental_volume")],
-    [Markup.button.callback("Модификаторы при разделении аудио", "set_audio_process_power")],
-    [Markup.button.callback("Включить реверберацию", "toggle_audio_reverb"), Markup.button.callback("Включить эхо", "toggle_audio_echo")],
-    [Markup.button.callback("Включить автотюн", "toggle_audio_autotune"),Markup.button.callback("Включить эффект телефона", "toggle_audio_phoneEffect")],
-    [Markup.button.callback("Настройка эффектов", "effects_settings")],
-    [Markup.button.callback("Меню", "menu")],
-]).resize();
-
-await ctx.replyWithMarkdown(settingsMessage, settingsKeyboard);
-
-  try{
-  await deletePreviousMessage(ctx);
-  }catch(err){
-    console.log(err)
-  }
-}
-
-
-export async function showEffectsSettings(ctx){
-
-  const session = ctx.session;
-
-  const settingsKeyboard = Markup.inlineKeyboard([
-    [Markup.button.callback("Эхо: Задержка - " + session.echoDelay, "effect_echo_delay"), Markup.button.callback("Эхо: Громкость - "+ session.echoPower, "effect_echo_power")],
-    [Markup.button.callback("Реверберация: Сила эффекта - "+session.reverbPower, "effect_reverb")],
-    [Markup.button.callback("Автотюн: Воздействие - "+session.autotune_attack, "effect_autotune_attack"), Markup.button.callback("Автотюн: Сила - "+session.autotune_strength, "effect_autotune_str")],
-    [Markup.button.callback("Назад", "aisettings")],
-  ]).resize();
-
-  await ctx.replyWithMarkdown("Меню настройки эффектов", settingsKeyboard);
-}
